+++ src/Data/Conduit/FSNotify.hs
@@ -3,6 +3,7 @@
 module Data.Conduit.FSNotify
     ( -- * Conduit API
       sourceFileChanges
+    , acquireSourceFileChanges
     , FileChangeSettings
     , mkFileChangeSettings
 
@@ -22,14 +23,15 @@ module Data.Conduit.FSNotify
 
 import Control.Exception (assert)
 import Data.Conduit
-import Control.Monad.Trans.Resource (MonadResource)
-import Control.Monad.IO.Class (liftIO)
+import Control.Monad.Trans.Resource (MonadResource, release)
+import Control.Monad.IO.Class (MonadIO, liftIO)
 import Control.Monad (forever)
 import qualified System.FSNotify as FS
 import System.Directory (canonicalizePath)
 import Control.Concurrent.Chan
 import Data.List (stripPrefix)
 import System.FilePath (addTrailingPathSeparator)
+import qualified Data.Acquire as A
 
 -- | Settings for watching for file changes, to be passed in to
 -- 'sourceFileChanges'. Should be created with 'mkFileChangeSettings'.
@@ -102,13 +104,30 @@ mkFileChangeSettings dir = FileChangeSettings
 sourceFileChanges :: MonadResource m
                   => FileChangeSettings
                   -> ConduitM i FS.Event m ()
-sourceFileChanges FileChangeSettings {..} =
-  -- The bracketP function allows us to safely allocate some resource
-  -- and guarantee it will be cleaned up. In our case, we are calling
-  -- startManager to allocate a file watching manager, and stopManager
-  -- to clean it up. These functions will under the surface tie in to
-  -- OS-specific file watch mechanisms, such as inotify on Linux.
-  bracketP (FS.startManagerConf fcsWatchConfig) FS.stopManager $ \man -> do
+sourceFileChanges fcs = do
+    (key, src) <- A.allocateAcquire $ acquireSourceFileChanges fcs
+    src
+    release key
+
+-- | The same as 'sourceFileChanges', but returned in an
+-- 'A.Acquire'. This is slightly clunkier to use than
+-- 'sourceFileChanges', but provides two benefits:
+--
+-- * It does not require 'MonadResource'
+--
+-- * You are guaranteed that the directory will be watched
+--   immediately. With 'sourceFileChanges', watching will only
+--   commence once you 'await' for the first change.
+--
+-- @since 0.1.1.0
+acquireSourceFileChanges
+  :: MonadIO m
+  => FileChangeSettings
+  -> A.Acquire (ConduitM i FS.Event m ())
+acquireSourceFileChanges FileChangeSettings {..} = do
+    -- Safely acquire a manager, guaranteeing it will be released
+    man <- A.mkAcquire (FS.startManagerConf fcsWatchConfig) FS.stopManager
+
     -- Get the absolute path of the root directory
     root' <- liftIO $ canonicalizePath fcsDir
 
@@ -120,7 +139,10 @@ sourceFileChanges FileChangeSettings {..} =
 
     -- Start watching a directory tree, accepting all events (const True).
     let watchChan = if fcsRecursive then FS.watchTreeChan else FS.watchDirChan
-    bracketP (watchChan man root' fcsPredicate chan) id $ const $ forever $ do
+
+    _ <- A.mkAcquire (watchChan man root' fcsPredicate chan) id
+
+    return $ forever $ do
         event <- liftIO $ readChan chan
 
         if fcsRelative