+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 * Drop older GHC support
 * Make test suite more reliable [#2](https://github.com/fpco/fsnotify-conduit/issues/2)
+* Provide `acquireSourceFileChanges`
 
 ## 0.1.0.0